BlockList: extract block edit functions as thunk actions
